refactor(upstream): extract resources_list.rs + resources_read.rs (lab-kvji.12.8 step 16)
- resources_list.rs: cached_upstream_resource_uris, gateway_servers_doc/_server_schema/
  _synthetic_resources, list_upstream_resources, subject_scoped_resources + normalize +
  4 gateway tests.
- resources_read.rs: read_upstream_resource + subject_scoped_read_resource +
  successful_resource_listing + read_resource_times_out tests.
NO-TOUCH verified: subject_scoped_read_resource retains subject threading + both
redact_resource_uri_for_logging call sites preserved (2 call sites, == baseline).
Recovered 2 tests accidentally dropped in earlier bulk deletes: empty_pool_has_no_tools
(→tools.rs), get_prompt_times_out (→pool.rs, moves to prompts_get in step 17). Test
count back to 1423. cached prompt-name/owner methods kept in pool.rs for step 17.

Bead: lab-kvji.12.8
