+++ crates/lab/src/dispatch/upstream/pool.rs
@@ -11,11 +11,7 @@ use futures::StreamExt;
 use futures::future::BoxFuture;
 use futures::stream::FuturesUnordered;
 use rmcp::RoleClient;
-use rmcp::model::{
-    AnnotateAble, GetPromptRequestParams, GetPromptResult, Prompt, RawResource, ReadResourceResult,
-    Resource,
-};
-use serde_json::Value;
+use rmcp::model::{GetPromptRequestParams, GetPromptResult, Prompt};
 use tokio::sync::{Mutex, RwLock};
 use tokio_util::sync::CancellationToken;
 
@@ -40,6 +36,8 @@ mod lifecycle;
 mod logging;
 mod probe;
 mod registration;
+mod resources_list;
+mod resources_read;
 #[cfg(test)]
 mod testsupport;
 mod tools;
@@ -53,7 +51,6 @@ pub use helpers::{UpstreamCachedSummary, in_process_upstream_name};
 // call sites unchanged while the bodies live in the child modules.
 use connect::*;
 use discover::*;
-use entries::*;
 use helpers::*;
 use logging::*;
 
@@ -163,513 +160,6 @@ impl UpstreamPool {
         self
     }
 
-    /// Return cached resource URIs keyed by upstream name (used in catalog snapshots).
-    pub async fn cached_upstream_resource_uris(&self) -> Vec<(String, Vec<String>)> {
-        let catalog = self.catalog.read().await;
-        catalog
-            .iter()
-            .filter(|(_, entry)| !entry.resource_uris.is_empty())
-            .map(|(name, entry)| (name.clone(), entry.resource_uris.clone()))
-            .collect()
-    }
-
-    /// Return cached prompt names from all upstreams, excluding any that clash with builtins.
-    pub async fn cached_upstream_prompt_names(&self, builtins: &[&str]) -> Vec<String> {
-        let catalog = self.catalog.read().await;
-        catalog
-            .values()
-            .flat_map(|entry| entry.prompt_names.iter().cloned())
-            .filter(|name| !builtins.contains(&name.as_str()))
-            .collect()
-    }
-
-    async fn cached_prompt_owner(
-        &self,
-        prompt_name: &str,
-        require_routable: bool,
-    ) -> Option<String> {
-        let catalog = self.catalog.read().await;
-        let mut entries = catalog.values().collect::<Vec<_>>();
-        entries.sort_unstable_by(|left, right| left.name.cmp(&right.name));
-        entries.into_iter().find_map(|entry| {
-            if require_routable && !entry.prompt_health.is_routable() {
-                return None;
-            }
-            entry
-                .prompt_names
-                .iter()
-                .any(|name| name == prompt_name)
-                .then(|| entry.name.to_string())
-        })
-    }
-
-    /// Render the synthetic `lab://gateway/servers` document.
-    ///
-    /// Lists every registered upstream (regardless of health) with the
-    /// tool count an agent would see in the corresponding schema document.
-    pub async fn gateway_servers_doc(&self) -> Value {
-        let catalog = self.catalog.read().await;
-        let mut servers: Vec<Value> = catalog
-            .iter()
-            .map(|(name, e)| {
-                let tool_count = e
-                    .tools
-                    .values()
-                    .filter(|t| e.exposure_policy.matches(&t.tool.name))
-                    .count();
-                serde_json::json!({
-                    "name": name,
-                    "tool_count": tool_count,
-                    "prompt_count": e.prompt_count,
-                    "resource_count": e.resource_count,
-                    "tool_health": health_str(e.tool_health),
-                    "tool_last_error": e.tool_last_error,
-                })
-            })
-            .collect();
-        servers.sort_by(|a, b| a["name"].as_str().cmp(&b["name"].as_str()));
-        serde_json::json!({ "servers": servers })
-    }
-
-    /// Render the synthetic `lab://gateway/<name>/schema` document.
-    ///
-    /// Returns `None` when the upstream is not registered. Tools hidden by
-    /// the upstream's `ToolExposurePolicy` are omitted. `input_schema` and
-    /// `meta` are passed through verbatim from the cached tool definition.
-    pub async fn gateway_server_schema(&self, name: &str) -> Option<Value> {
-        let catalog = self.catalog.read().await;
-        let entry = catalog.get(name)?;
-        let mut tools: Vec<Value> = entry
-            .tools
-            .values()
-            .filter(|t| entry.exposure_policy.matches(&t.tool.name))
-            .map(|t| {
-                serde_json::json!({
-                    "name": t.tool.name.as_ref(),
-                    "description": t.tool.description.as_ref().map(|s| s.as_ref()),
-                    "input_schema": t.input_schema,
-                    "meta": t.tool.meta,
-                })
-            })
-            .collect();
-        tools.sort_by(|a, b| a["name"].as_str().cmp(&b["name"].as_str()));
-        Some(serde_json::json!({
-            "name": name,
-            "tools": tools,
-            "health": health_str(entry.tool_health),
-            "last_error": entry.tool_last_error,
-        }))
-    }
-
-    /// Synthetic gateway resources to emit from `list_resources`.
-    ///
-    /// Returns one entry for `lab://gateway/servers` plus one
-    /// `lab://gateway/<name>/schema` entry per registered upstream.
-    pub async fn gateway_synthetic_resources(&self) -> Vec<Resource> {
-        let mut out = vec![
-            RawResource::new("lab://gateway/servers", "gateway/servers")
-                .with_description("Index of upstream MCP servers registered with the gateway")
-                .with_mime_type("application/json")
-                .no_annotation(),
-        ];
-        let catalog = self.catalog.read().await;
-        let mut names: Vec<&String> = catalog.keys().collect();
-        names.sort();
-        for name in names {
-            out.push(
-                RawResource::new(
-                    format!("lab://gateway/{name}/schema"),
-                    format!("gateway/{name}/schema"),
-                )
-                .with_description(format!("Tool schemas for upstream `{name}`"))
-                .with_mime_type("application/json")
-                .no_annotation(),
-            );
-        }
-        out
-    }
-
-    /// List resources from all resource-proxy-enabled upstreams.
-    ///
-    /// Resources are prefixed with `lab://upstream/{name}/` to avoid collisions.
-    pub async fn list_upstream_resources(&self) -> Vec<Resource> {
-        let peers = routable_upstream_peers(self, UpstreamCapability::Resources).await;
-        if peers.is_empty() {
-            return Vec::new();
-        }
-
-        // Issue RPCs in parallel, then sort by upstream name for deterministic order.
-        let mut futures = FuturesUnordered::new();
-        for (name, peer) in peers {
-            futures.push(async move {
-                let result = peer.list_resources(None).await;
-                (name, result)
-            });
-        }
-
-        let mut results: Vec<(String, Result<_, _>)> = Vec::new();
-        while let Some(item) = futures.next().await {
-            results.push(item);
-        }
-        results.sort_unstable_by(|a, b| a.0.cmp(&b.0));
-
-        let mut resources = Vec::new();
-        for (name, result) in results {
-            match result {
-                Ok(result) => {
-                    self.record_success_for(&name, UpstreamCapability::Resources)
-                        .await;
-                    let resource_uris = result
-                        .resources
-                        .iter()
-                        .map(|resource| bare_upstream_resource_uri(&resource.uri).to_string())
-                        .collect();
-                    {
-                        let mut catalog = self.catalog.write().await;
-                        if let Some(entry) = catalog.get_mut(&name) {
-                            entry.resource_count = result.resources.len();
-                            entry.resource_uris = resource_uris;
-                        }
-                    }
-                    for mut resource in result.resources {
-                        rewrite_resource_uri(&mut resource, &name);
-                        resources.push(resource);
-                    }
-                }
-                Err(e) => {
-                    self.record_failure_for(
-                        &name,
-                        UpstreamCapability::Resources,
-                        format!("failed to list resources from upstream: {e}"),
-                    )
-                    .await;
-                    {
-                        let mut catalog = self.catalog.write().await;
-                        if let Some(entry) = catalog.get_mut(&name) {
-                            entry.resource_count = 0;
-                            entry.resource_uris.clear();
-                        }
-                    }
-                    tracing::warn!(
-                        upstream = %name,
-                        error = %e,
-                        "failed to list resources from upstream"
-                    );
-                }
-            }
-        }
-
-        resources
-    }
-
-    pub async fn subject_scoped_resources(
-        &self,
-        configs: &[UpstreamConfig],
-        subject: &str,
-    ) -> Vec<Resource> {
-        let mut futures = FuturesUnordered::new();
-        let oauth_client_cache = self.oauth_client_cache.clone();
-        for config in configs
-            .iter()
-            .filter(|config| config.oauth.is_some() && config.proxy_resources)
-        {
-            let config = config.clone();
-            let subject = subject.to_string();
-            let oauth_client_cache = oauth_client_cache.clone();
-            futures.push(async move {
-                let result = connect_upstream(
-                    &config,
-                    Some(subject.as_str()),
-                    oauth_client_cache.as_ref(),
-                    None,
-                    None,
-                )
-                .await
-                .map(|(conn, _)| conn);
-                (config.name.clone(), result)
-            });
-        }
-
-        let mut resources = Vec::new();
-        while let Some((name, result)) = futures.next().await {
-            let Ok(conn) = result else {
-                continue;
-            };
-            match conn.peer.list_resources(None).await {
-                Ok(result) => {
-                    for mut resource in result.resources {
-                        rewrite_resource_uri(&mut resource, &name);
-                        resources.push(resource);
-                    }
-                }
-                Err(error) => {
-                    tracing::warn!(
-                        upstream = %name,
-                        error = %error,
-                        "subject-scoped upstream resource discovery failed"
-                    );
-                }
-            }
-        }
-
-        resources
-    }
-
-    /// Read a resource from an upstream, given a prefixed URI.
-    ///
-    /// Expects URIs in the form `lab://upstream/{name}/{original_uri}`.
-    /// Returns `None` if the upstream name is not found or not resource-enabled.
-    pub async fn read_upstream_resource(
-        &self,
-        uri: &str,
-    ) -> Option<Result<ReadResourceResult, String>> {
-        let start = Instant::now();
-        let prefix = "lab://upstream/";
-        let rest = uri.strip_prefix(prefix)?;
-
-        // Extract upstream name and original URI
-        let slash_pos = rest.find('/')?;
-        let upstream_name = &rest[..slash_pos];
-        let original_uri = &rest[slash_pos + 1..];
-
-        // Check if this upstream has resource proxying enabled.
-        // Clone the vec and drop the lock before any async work.
-        let is_resource_enabled = {
-            let resource_names = self.resource_upstreams.read().await;
-            if !resource_names.iter().any(|n| n == upstream_name) {
-                false
-            } else {
-                let catalog = self.catalog.read().await;
-                catalog
-                    .get(upstream_name)
-                    .is_some_and(|entry| entry.resource_health.is_routable())
-            }
-        };
-        if !is_resource_enabled {
-            return None;
-        }
-
-        // Clone the peer handle out, then drop the lock before awaiting.
-        let peer = self
-            .acquire_peer(
-                upstream_name,
-                UpstreamCapability::Resources,
-                "resource.read",
-            )
-            .await?;
-
-        let redacted_uri = redact_resource_uri_for_logging(uri);
-        let event = UpstreamRequestLog::resource(upstream_name, redacted_uri, false);
-        log_upstream_request_start(event);
-
-        let params = rmcp::model::ReadResourceRequestParams::new(original_uri);
-
-        let result =
-            match tokio::time::timeout(self.request_timeout, peer.read_resource(params)).await {
-                Ok(Ok(result)) => {
-                    let normalized = normalize_resource_result_uri(result, uri);
-                    Ok(normalized)
-                }
-                Ok(Err(e)) => {
-                    self.record_failure_for(
-                        upstream_name,
-                        UpstreamCapability::Resources,
-                        format!("upstream resource read failed: {e}"),
-                    )
-                    .await;
-                    log_upstream_request_error(
-                        event,
-                        start.elapsed().as_millis(),
-                        "upstream_error",
-                        Some(&e),
-                        None,
-                        None,
-                    );
-                    Err(format!("upstream resource read failed: {e}"))
-                }
-                Err(_) => {
-                    let message = format!(
-                        "upstream resource read timed out after {}ms",
-                        self.request_timeout.as_millis()
-                    );
-                    self.record_failure_for(
-                        upstream_name,
-                        UpstreamCapability::Resources,
-                        message.clone(),
-                    )
-                    .await;
-                    log_upstream_request_error(
-                        event,
-                        start.elapsed().as_millis(),
-                        "timeout",
-                        None,
-                        None,
-                        None,
-                    );
-                    Err(message)
-                }
-            };
-
-        // Enforce the same response size cap as call_tool (post-hoc).
-        if let Ok(ref r) = result {
-            let response_size = serde_json::to_string(r).map_or(0, |s| s.len());
-            let max_bytes = max_response_bytes();
-            if response_size > max_bytes {
-                self.record_failure_for(
-                    upstream_name,
-                    UpstreamCapability::Resources,
-                    format!(
-                        "upstream resource response too large ({response_size} bytes, max {max_bytes})"
-                    ),
-                )
-                .await;
-                log_upstream_request_error(
-                    event,
-                    start.elapsed().as_millis(),
-                    "response_too_large",
-                    None,
-                    Some(response_size),
-                    Some(max_bytes),
-                );
-                return Some(Err(format!(
-                    "upstream resource response too large ({response_size} bytes, max {max_bytes})"
-                )));
-            }
-            self.record_success_for(upstream_name, UpstreamCapability::Resources)
-                .await;
-            log_upstream_request_finish(event, start.elapsed().as_millis(), Some(response_size));
-        }
-
-        Some(result)
-    }
-
-    pub async fn subject_scoped_read_resource(
-        &self,
-        config: &UpstreamConfig,
-        subject: &str,
-        uri: &str,
-    ) -> Result<ReadResourceResult, String> {
-        let start = Instant::now();
-        let prefix = format!("lab://upstream/{}/", config.name);
-        let original_uri = uri
-            .strip_prefix(&prefix)
-            .ok_or_else(|| "resource uri does not match upstream".to_string())?;
-        let redacted_uri = redact_resource_uri_for_logging(uri);
-        let event = UpstreamRequestLog::resource(&config.name, redacted_uri, true)
-            .with_transport(upstream_transport(config));
-        log_upstream_request_start(event);
-        let (conn, _) = match connect_upstream(
-            config,
-            Some(subject),
-            self.oauth_client_cache.as_ref(),
-            None,
-            None,
-        )
-        .await
-        {
-            Ok(conn) => conn,
-            Err(error) => {
-                self.record_failure_for(
-                    &config.name,
-                    UpstreamCapability::Resources,
-                    format!("upstream resource connect failed: {error}"),
-                )
-                .await;
-                log_upstream_request_error(
-                    event,
-                    start.elapsed().as_millis(),
-                    "upstream_connect_error",
-                    Some(&error),
-                    None,
-                    None,
-                );
-                return Err(error.to_string());
-            }
-        };
-        match tokio::time::timeout(
-            self.request_timeout,
-            conn.peer
-                .read_resource(rmcp::model::ReadResourceRequestParams::new(original_uri)),
-        )
-        .await
-        {
-            Ok(Ok(result)) => {
-                // Size check before recording success so an oversized response
-                // does not advance the circuit breaker's healthy counter.
-                let response_size = serde_json::to_string(&result).map_or(0, |s| s.len());
-                let max_bytes = max_response_bytes();
-                if response_size > max_bytes {
-                    self.record_failure_for(
-                        &config.name,
-                        UpstreamCapability::Resources,
-                        format!(
-                            "upstream resource response too large ({response_size} bytes, max {max_bytes})"
-                        ),
-                    )
-                    .await;
-                    log_upstream_request_error(
-                        event,
-                        start.elapsed().as_millis(),
-                        "response_too_large",
-                        None,
-                        Some(response_size),
-                        Some(max_bytes),
-                    );
-                    return Err(format!(
-                        "upstream resource response too large ({response_size} bytes, max {max_bytes})"
-                    ));
-                }
-                self.record_success_for(&config.name, UpstreamCapability::Resources)
-                    .await;
-                let normalized = normalize_resource_result_uri(result, uri);
-                log_upstream_request_finish(
-                    event,
-                    start.elapsed().as_millis(),
-                    Some(response_size),
-                );
-                Ok(normalized)
-            }
-            Ok(Err(error)) => {
-                self.record_failure_for(
-                    &config.name,
-                    UpstreamCapability::Resources,
-                    format!("upstream resource read failed: {error}"),
-                )
-                .await;
-                log_upstream_request_error(
-                    event,
-                    start.elapsed().as_millis(),
-                    "upstream_error",
-                    Some(&error),
-                    None,
-                    None,
-                );
-                Err(format!("upstream resource read failed: {error}"))
-            }
-            Err(_) => {
-                let message = format!(
-                    "upstream resource read timed out after {}ms",
-                    self.request_timeout.as_millis()
-                );
-                self.record_failure_for(
-                    &config.name,
-                    UpstreamCapability::Resources,
-                    message.clone(),
-                )
-                .await;
-                log_upstream_request_error(
-                    event,
-                    start.elapsed().as_millis(),
-                    "timeout",
-                    None,
-                    None,
-                    None,
-                );
-                Err(message)
-            }
-        }
-    }
-
     /// Fetch prompts from all healthy upstreams and merge them, returning both the
     /// deduplicated prompt list and the ownership map (prompt_name -> upstream_name).
     ///
@@ -752,6 +242,36 @@ impl UpstreamPool {
     pub async fn list_upstream_prompts(&self, builtin_names: &[&str]) -> Vec<Prompt> {
         let (prompts, _) = self.collect_upstream_prompts(builtin_names).await;
         prompts
+    }
+
+    /// Return cached prompt names from all upstreams, excluding any that clash with builtins.
+    pub async fn cached_upstream_prompt_names(&self, builtins: &[&str]) -> Vec<String> {
+        let catalog = self.catalog.read().await;
+        catalog
+            .values()
+            .flat_map(|entry| entry.prompt_names.iter().cloned())
+            .filter(|name| !builtins.contains(&name.as_str()))
+            .collect()
+    }
+
+    async fn cached_prompt_owner(
+        &self,
+        prompt_name: &str,
+        require_routable: bool,
+    ) -> Option<String> {
+        let catalog = self.catalog.read().await;
+        let mut entries = catalog.values().collect::<Vec<_>>();
+        entries.sort_unstable_by(|left, right| left.name.cmp(&right.name));
+        entries.into_iter().find_map(|entry| {
+            if require_routable && !entry.prompt_health.is_routable() {
+                return None;
+            }
+            entry
+                .prompt_names
+                .iter()
+                .any(|name| name == prompt_name)
+                .then(|| entry.name.to_string())
+        })
     }
 
     pub async fn subject_scoped_prompts(
@@ -1026,13 +546,12 @@ impl Default for UpstreamPool {
 #[cfg(test)]
 mod tests {
     use super::super::types;
-    use super::super::types::{ToolExposurePolicy, UpstreamTool};
     use super::testsupport::*;
     use super::*;
     use crate::mcp::logging::logging_level_rank;
     use crate::mcp::server::LabMcpServer;
     use crate::registry::ToolRegistry;
-    use rmcp::model::{LoggingLevel, ResourceContents};
+    use rmcp::model::LoggingLevel;
     use std::sync::atomic::{AtomicU8, Ordering};
 
     #[test]
@@ -1055,123 +574,6 @@ mod tests {
         assert_eq!(owners.get("shared").map(String::as_str), Some("alpha"));
         assert_eq!(owners.get("left-only").map(String::as_str), Some("alpha"));
         assert_eq!(owners.get("right-only").map(String::as_str), Some("zeta"));
-    }
-
-    #[test]
-    fn normalize_resource_result_uri_rewrites_all_contents() {
-        let result = ReadResourceResult::new(vec![
-            ResourceContents::text("hello", "http://upstream/resource"),
-            ResourceContents::blob("YWJj", "file:///tmp/upstream"),
-        ]);
-
-        let normalized =
-            normalize_resource_result_uri(result, "lab://upstream/demo/http://upstream/resource");
-
-        let uris: Vec<_> = normalized
-            .contents
-            .iter()
-            .map(|content| match content {
-                ResourceContents::TextResourceContents { uri, .. }
-                | ResourceContents::BlobResourceContents { uri, .. } => uri.as_str(),
-            })
-            .collect();
-
-        assert_eq!(
-            uris,
-            vec![
-                "lab://upstream/demo/http://upstream/resource",
-                "lab://upstream/demo/http://upstream/resource",
-            ]
-        );
-    }
-
-    #[tokio::test]
-    async fn empty_pool_has_no_tools() {
-        let pool = UpstreamPool::new();
-        assert!(pool.healthy_tools().await.is_empty());
-        assert_eq!(pool.upstream_count().await, 0);
-    }
-
-    #[tokio::test]
-    async fn read_resource_times_out_slow_upstream_response() {
-        let pool = slow_response_pool("slow").await;
-
-        let result = pool
-            .read_upstream_resource("lab://upstream/slow/file:///tmp/slow")
-            .await
-            .expect("resource upstream is enabled")
-            .expect_err("slow resource read should time out");
-
-        assert!(result.contains("timed out"));
-    }
-
-    #[tokio::test]
-    async fn get_prompt_times_out_slow_upstream_response() {
-        let pool = slow_response_pool("slow").await;
-
-        let result = pool
-            .get_prompt("slow", GetPromptRequestParams::new("slow.prompt"))
-            .await
-            .expect("upstream is connected")
-            .expect_err("slow prompt get should time out");
-
-        assert!(result.contains("timed out"));
-    }
-
-    #[tokio::test]
-    async fn successful_resource_listing_populates_snapshot_cache() {
-        let pool = static_catalog_pool("static").await;
-
-        let resources = pool.list_upstream_resources().await;
-        let listed_uris: Vec<_> = resources
-            .iter()
-            .map(|resource| resource.uri.as_str())
-            .collect();
-        assert_eq!(
-            listed_uris,
-            vec![
-                "lab://upstream/static/file:///tmp/upstream-one",
-                "lab://upstream/static/file:///tmp/upstream-two",
-            ]
-        );
-
-        let cached = pool.cached_upstream_resource_uris().await;
-        assert_eq!(
-            cached,
-            vec![(
-                "static".to_string(),
-                vec![
-                    "file:///tmp/upstream-one".to_string(),
-                    "file:///tmp/upstream-two".to_string(),
-                ],
-            )]
-        );
-
-        let runtime = crate::dispatch::gateway::manager::GatewayRuntimeHandle::default();
-        runtime.swap(Some(Arc::clone(&pool))).await;
-        let manager = Arc::new(crate::dispatch::gateway::manager::GatewayManager::new(
-            std::path::PathBuf::from("config.toml"),
-            runtime,
-        ));
-        let server = LabMcpServer {
-            registry: Arc::new(ToolRegistry::new()),
-            gateway_manager: Some(manager),
-            node_role: None,
-            peers: Arc::new(RwLock::new(Vec::new())),
-            logging_level: Arc::new(AtomicU8::new(logging_level_rank(LoggingLevel::Info))),
-        };
-
-        let snapshot = server.snapshot_catalog().await;
-        assert!(
-            snapshot
-                .resources
-                .contains("lab://upstream/static/file:///tmp/upstream-one")
-        );
-        assert!(
-            snapshot
-                .resources
-                .contains("lab://upstream/static/file:///tmp/upstream-two")
-        );
     }
 
     #[tokio::test]
@@ -1270,109 +672,15 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn gateway_servers_doc_lists_one_healthy_upstream() {
-        use std::sync::Arc;
+    async fn get_prompt_times_out_slow_upstream_response() {
+        let pool = slow_response_pool("slow").await;
 
-        let pool = UpstreamPool::new();
-        let mut tools = HashMap::new();
-        tools.insert(
-            "search".to_string(),
-            UpstreamTool {
-                tool: rmcp::model::Tool::new(
-                    "search",
-                    "search the index",
-                    Arc::new(serde_json::Map::new()),
-                ),
-                input_schema: Some(serde_json::json!({"type": "object"})),
-                output_schema: None,
-                upstream_name: Arc::from("alpha"),
-                destructive: false,
-            },
-        );
-        let entry = healthy_in_process_entry(Arc::from("alpha"), tools);
-        pool.catalog
-            .write()
+        let result = pool
+            .get_prompt("slow", GetPromptRequestParams::new("slow.prompt"))
             .await
-            .insert("alpha".to_string(), entry);
+            .expect("upstream is connected")
+            .expect_err("slow prompt get should time out");
 
-        let doc = pool.gateway_servers_doc().await;
-        let servers = doc
-            .get("servers")
-            .and_then(|v| v.as_array())
-            .expect("servers array");
-        assert_eq!(servers.len(), 1);
-        let s = &servers[0];
-        assert_eq!(s["name"], "alpha");
-        assert_eq!(s["tool_count"], 1);
-        assert_eq!(s["tool_health"], "healthy");
-        assert!(s["tool_last_error"].is_null());
-        assert_eq!(s["prompt_count"], 0);
-        assert_eq!(s["resource_count"], 0);
-    }
-
-    #[tokio::test]
-    async fn gateway_server_schema_respects_exposure_policy() {
-        use std::sync::Arc;
-
-        let make_tool = |name: &'static str| UpstreamTool {
-            tool: rmcp::model::Tool::new(name, "desc", Arc::new(serde_json::Map::new())),
-            input_schema: Some(serde_json::json!({"type": "object"})),
-            output_schema: None,
-            upstream_name: Arc::from("alpha"),
-            destructive: false,
-        };
-
-        let mut tools = HashMap::new();
-        tools.insert("github_create".into(), make_tool("github_create"));
-        tools.insert("delete_repo".into(), make_tool("delete_repo"));
-
-        let mut entry = healthy_in_process_entry(Arc::from("alpha"), tools);
-        entry.exposure_policy =
-            ToolExposurePolicy::from_patterns(vec!["github_*".into()]).expect("policy");
-
-        let pool = UpstreamPool::new();
-        pool.catalog
-            .write()
-            .await
-            .insert("alpha".to_string(), entry);
-
-        let doc = pool.gateway_server_schema("alpha").await.expect("doc");
-        let names: Vec<&str> = doc["tools"]
-            .as_array()
-            .expect("tools array")
-            .iter()
-            .map(|t| t["name"].as_str().expect("name"))
-            .collect();
-        assert_eq!(names, vec!["github_create"]);
-        assert_eq!(doc["health"], "healthy");
-        assert!(doc["last_error"].is_null());
-        assert_eq!(doc["name"], "alpha");
-    }
-
-    #[tokio::test]
-    async fn gateway_server_schema_unknown_upstream_returns_none() {
-        let pool = UpstreamPool::new();
-        assert!(pool.gateway_server_schema("nope").await.is_none());
-    }
-
-    #[tokio::test]
-    async fn gateway_synthetic_resources_lists_index_and_per_upstream() {
-        use std::sync::Arc;
-
-        let pool = UpstreamPool::new();
-        let entry = healthy_in_process_entry(Arc::from("alpha"), HashMap::new());
-        pool.catalog
-            .write()
-            .await
-            .insert("alpha".to_string(), entry);
-        let entry = healthy_in_process_entry(Arc::from("beta"), HashMap::new());
-        pool.catalog.write().await.insert("beta".to_string(), entry);
-
-        let resources = pool.gateway_synthetic_resources().await;
-        let uris: Vec<String> = resources.iter().map(|r| r.uri.clone()).collect();
-        assert!(uris.iter().any(|u| u == "lab://gateway/servers"));
-        assert!(uris.iter().any(|u| u == "lab://gateway/alpha/schema"));
-        assert!(uris.iter().any(|u| u == "lab://gateway/beta/schema"));
-        assert_eq!(uris.len(), 3);
+        assert!(result.contains("timed out"));
     }
 }
