+++ crates/lab/src/dispatch/upstream/pool/resources_list.rs
@@ -1,0 +1,391 @@
+//! Resource listing and the synthetic gateway documents.
+//!
+//! `list_upstream_resources` / `subject_scoped_resources` enumerate proxied
+//! upstream resources (rewriting URIs to the gateway-prefixed form), while the
+//! `gateway_*` methods render the synthetic `lab://gateway/*` documents and
+//! resources. `cached_upstream_resource_uris` exposes the cached snapshot.
+
+use futures::StreamExt;
+use futures::stream::FuturesUnordered;
+use rmcp::model::{AnnotateAble, RawResource, Resource};
+use serde_json::Value;
+
+use crate::config::UpstreamConfig;
+
+use super::super::types::UpstreamCapability;
+use super::UpstreamPool;
+use super::connect::connect_upstream;
+use super::discover::routable_upstream_peers;
+use super::entries::health_str;
+use super::helpers::{bare_upstream_resource_uri, rewrite_resource_uri};
+
+impl UpstreamPool {
+    /// Return cached resource URIs keyed by upstream name (used in catalog snapshots).
+    pub async fn cached_upstream_resource_uris(&self) -> Vec<(String, Vec<String>)> {
+        let catalog = self.catalog.read().await;
+        catalog
+            .iter()
+            .filter(|(_, entry)| !entry.resource_uris.is_empty())
+            .map(|(name, entry)| (name.clone(), entry.resource_uris.clone()))
+            .collect()
+    }
+
+    /// Render the synthetic `lab://gateway/servers` document.
+    ///
+    /// Lists every registered upstream (regardless of health) with the
+    /// tool count an agent would see in the corresponding schema document.
+    pub async fn gateway_servers_doc(&self) -> Value {
+        let catalog = self.catalog.read().await;
+        let mut servers: Vec<Value> = catalog
+            .iter()
+            .map(|(name, e)| {
+                let tool_count = e
+                    .tools
+                    .values()
+                    .filter(|t| e.exposure_policy.matches(&t.tool.name))
+                    .count();
+                serde_json::json!({
+                    "name": name,
+                    "tool_count": tool_count,
+                    "prompt_count": e.prompt_count,
+                    "resource_count": e.resource_count,
+                    "tool_health": health_str(e.tool_health),
+                    "tool_last_error": e.tool_last_error,
+                })
+            })
+            .collect();
+        servers.sort_by(|a, b| a["name"].as_str().cmp(&b["name"].as_str()));
+        serde_json::json!({ "servers": servers })
+    }
+
+    /// Render the synthetic `lab://gateway/<name>/schema` document.
+    ///
+    /// Returns `None` when the upstream is not registered. Tools hidden by
+    /// the upstream's `ToolExposurePolicy` are omitted. `input_schema` and
+    /// `meta` are passed through verbatim from the cached tool definition.
+    pub async fn gateway_server_schema(&self, name: &str) -> Option<Value> {
+        let catalog = self.catalog.read().await;
+        let entry = catalog.get(name)?;
+        let mut tools: Vec<Value> = entry
+            .tools
+            .values()
+            .filter(|t| entry.exposure_policy.matches(&t.tool.name))
+            .map(|t| {
+                serde_json::json!({
+                    "name": t.tool.name.as_ref(),
+                    "description": t.tool.description.as_ref().map(|s| s.as_ref()),
+                    "input_schema": t.input_schema,
+                    "meta": t.tool.meta,
+                })
+            })
+            .collect();
+        tools.sort_by(|a, b| a["name"].as_str().cmp(&b["name"].as_str()));
+        Some(serde_json::json!({
+            "name": name,
+            "tools": tools,
+            "health": health_str(entry.tool_health),
+            "last_error": entry.tool_last_error,
+        }))
+    }
+
+    /// Synthetic gateway resources to emit from `list_resources`.
+    ///
+    /// Returns one entry for `lab://gateway/servers` plus one
+    /// `lab://gateway/<name>/schema` entry per registered upstream.
+    pub async fn gateway_synthetic_resources(&self) -> Vec<Resource> {
+        let mut out = vec![
+            RawResource::new("lab://gateway/servers", "gateway/servers")
+                .with_description("Index of upstream MCP servers registered with the gateway")
+                .with_mime_type("application/json")
+                .no_annotation(),
+        ];
+        let catalog = self.catalog.read().await;
+        let mut names: Vec<&String> = catalog.keys().collect();
+        names.sort();
+        for name in names {
+            out.push(
+                RawResource::new(
+                    format!("lab://gateway/{name}/schema"),
+                    format!("gateway/{name}/schema"),
+                )
+                .with_description(format!("Tool schemas for upstream `{name}`"))
+                .with_mime_type("application/json")
+                .no_annotation(),
+            );
+        }
+        out
+    }
+
+    /// List resources from all resource-proxy-enabled upstreams.
+    ///
+    /// Resources are prefixed with `lab://upstream/{name}/` to avoid collisions.
+    pub async fn list_upstream_resources(&self) -> Vec<Resource> {
+        let peers = routable_upstream_peers(self, UpstreamCapability::Resources).await;
+        if peers.is_empty() {
+            return Vec::new();
+        }
+
+        // Issue RPCs in parallel, then sort by upstream name for deterministic order.
+        let mut futures = FuturesUnordered::new();
+        for (name, peer) in peers {
+            futures.push(async move {
+                let result = peer.list_resources(None).await;
+                (name, result)
+            });
+        }
+
+        let mut results: Vec<(String, Result<_, _>)> = Vec::new();
+        while let Some(item) = futures.next().await {
+            results.push(item);
+        }
+        results.sort_unstable_by(|a, b| a.0.cmp(&b.0));
+
+        let mut resources = Vec::new();
+        for (name, result) in results {
+            match result {
+                Ok(result) => {
+                    self.record_success_for(&name, UpstreamCapability::Resources)
+                        .await;
+                    let resource_uris = result
+                        .resources
+                        .iter()
+                        .map(|resource| bare_upstream_resource_uri(&resource.uri).to_string())
+                        .collect();
+                    {
+                        let mut catalog = self.catalog.write().await;
+                        if let Some(entry) = catalog.get_mut(&name) {
+                            entry.resource_count = result.resources.len();
+                            entry.resource_uris = resource_uris;
+                        }
+                    }
+                    for mut resource in result.resources {
+                        rewrite_resource_uri(&mut resource, &name);
+                        resources.push(resource);
+                    }
+                }
+                Err(e) => {
+                    self.record_failure_for(
+                        &name,
+                        UpstreamCapability::Resources,
+                        format!("failed to list resources from upstream: {e}"),
+                    )
+                    .await;
+                    {
+                        let mut catalog = self.catalog.write().await;
+                        if let Some(entry) = catalog.get_mut(&name) {
+                            entry.resource_count = 0;
+                            entry.resource_uris.clear();
+                        }
+                    }
+                    tracing::warn!(
+                        upstream = %name,
+                        error = %e,
+                        "failed to list resources from upstream"
+                    );
+                }
+            }
+        }
+
+        resources
+    }
+
+    pub async fn subject_scoped_resources(
+        &self,
+        configs: &[UpstreamConfig],
+        subject: &str,
+    ) -> Vec<Resource> {
+        let mut futures = FuturesUnordered::new();
+        let oauth_client_cache = self.oauth_client_cache.clone();
+        for config in configs
+            .iter()
+            .filter(|config| config.oauth.is_some() && config.proxy_resources)
+        {
+            let config = config.clone();
+            let subject = subject.to_string();
+            let oauth_client_cache = oauth_client_cache.clone();
+            futures.push(async move {
+                let result = connect_upstream(
+                    &config,
+                    Some(subject.as_str()),
+                    oauth_client_cache.as_ref(),
+                    None,
+                    None,
+                )
+                .await
+                .map(|(conn, _)| conn);
+                (config.name.clone(), result)
+            });
+        }
+
+        let mut resources = Vec::new();
+        while let Some((name, result)) = futures.next().await {
+            let Ok(conn) = result else {
+                continue;
+            };
+            match conn.peer.list_resources(None).await {
+                Ok(result) => {
+                    for mut resource in result.resources {
+                        rewrite_resource_uri(&mut resource, &name);
+                        resources.push(resource);
+                    }
+                }
+                Err(error) => {
+                    tracing::warn!(
+                        upstream = %name,
+                        error = %error,
+                        "subject-scoped upstream resource discovery failed"
+                    );
+                }
+            }
+        }
+
+        resources
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+
+    use rmcp::model::{ReadResourceResult, ResourceContents};
+
+    use super::super::super::types::{ToolExposurePolicy, UpstreamTool};
+    use super::super::entries::healthy_in_process_entry;
+    use super::super::helpers::normalize_resource_result_uri;
+    use super::*;
+
+    #[test]
+    fn normalize_resource_result_uri_rewrites_all_contents() {
+        let result = ReadResourceResult::new(vec![
+            ResourceContents::text("hello", "http://upstream/resource"),
+            ResourceContents::blob("YWJj", "file:///tmp/upstream"),
+        ]);
+
+        let normalized =
+            normalize_resource_result_uri(result, "lab://upstream/demo/http://upstream/resource");
+
+        let uris: Vec<_> = normalized
+            .contents
+            .iter()
+            .map(|content| match content {
+                ResourceContents::TextResourceContents { uri, .. }
+                | ResourceContents::BlobResourceContents { uri, .. } => uri.as_str(),
+            })
+            .collect();
+
+        assert_eq!(
+            uris,
+            vec![
+                "lab://upstream/demo/http://upstream/resource",
+                "lab://upstream/demo/http://upstream/resource",
+            ]
+        );
+    }
+
+    #[tokio::test]
+    async fn gateway_servers_doc_lists_one_healthy_upstream() {
+        use std::sync::Arc;
+
+        let pool = UpstreamPool::new();
+        let mut tools = HashMap::new();
+        tools.insert(
+            "search".to_string(),
+            UpstreamTool {
+                tool: rmcp::model::Tool::new(
+                    "search",
+                    "search the index",
+                    Arc::new(serde_json::Map::new()),
+                ),
+                input_schema: Some(serde_json::json!({"type": "object"})),
+                output_schema: None,
+                upstream_name: Arc::from("alpha"),
+                destructive: false,
+            },
+        );
+        let entry = healthy_in_process_entry(Arc::from("alpha"), tools);
+        pool.catalog
+            .write()
+            .await
+            .insert("alpha".to_string(), entry);
+
+        let doc = pool.gateway_servers_doc().await;
+        let servers = doc
+            .get("servers")
+            .and_then(|v| v.as_array())
+            .expect("servers array");
+        assert_eq!(servers.len(), 1);
+        let s = &servers[0];
+        assert_eq!(s["name"], "alpha");
+        assert_eq!(s["tool_count"], 1);
+        assert_eq!(s["tool_health"], "healthy");
+        assert!(s["tool_last_error"].is_null());
+        assert_eq!(s["prompt_count"], 0);
+        assert_eq!(s["resource_count"], 0);
+    }
+
+    #[tokio::test]
+    async fn gateway_server_schema_respects_exposure_policy() {
+        use std::sync::Arc;
+
+        let make_tool = |name: &'static str| UpstreamTool {
+            tool: rmcp::model::Tool::new(name, "desc", Arc::new(serde_json::Map::new())),
+            input_schema: Some(serde_json::json!({"type": "object"})),
+            output_schema: None,
+            upstream_name: Arc::from("alpha"),
+            destructive: false,
+        };
+
+        let mut tools = HashMap::new();
+        tools.insert("github_create".into(), make_tool("github_create"));
+        tools.insert("delete_repo".into(), make_tool("delete_repo"));
+
+        let mut entry = healthy_in_process_entry(Arc::from("alpha"), tools);
+        entry.exposure_policy =
+            ToolExposurePolicy::from_patterns(vec!["github_*".into()]).expect("policy");
+
+        let pool = UpstreamPool::new();
+        pool.catalog
+            .write()
+            .await
+            .insert("alpha".to_string(), entry);
+
+        let doc = pool.gateway_server_schema("alpha").await.expect("doc");
+        let names: Vec<&str> = doc["tools"]
+            .as_array()
+            .expect("tools array")
+            .iter()
+            .map(|t| t["name"].as_str().expect("name"))
+            .collect();
+        assert_eq!(names, vec!["github_create"]);
+        assert_eq!(doc["health"], "healthy");
+        assert!(doc["last_error"].is_null());
+        assert_eq!(doc["name"], "alpha");
+    }
+
+    #[tokio::test]
+    async fn gateway_server_schema_unknown_upstream_returns_none() {
+        let pool = UpstreamPool::new();
+        assert!(pool.gateway_server_schema("nope").await.is_none());
+    }
+
+    #[tokio::test]
+    async fn gateway_synthetic_resources_lists_index_and_per_upstream() {
+        use std::sync::Arc;
+
+        let pool = UpstreamPool::new();
+        let entry = healthy_in_process_entry(Arc::from("alpha"), HashMap::new());
+        pool.catalog
+            .write()
+            .await
+            .insert("alpha".to_string(), entry);
+        let entry = healthy_in_process_entry(Arc::from("beta"), HashMap::new());
+        pool.catalog.write().await.insert("beta".to_string(), entry);
+
+        let resources = pool.gateway_synthetic_resources().await;
+        let uris: Vec<String> = resources.iter().map(|r| r.uri.clone()).collect();
+        assert!(uris.iter().any(|u| u == "lab://gateway/servers"));
+        assert!(uris.iter().any(|u| u == "lab://gateway/alpha/schema"));
+        assert!(uris.iter().any(|u| u == "lab://gateway/beta/schema"));
+        assert_eq!(uris.len(), 3);
+    }
+}
